requirements.txt
jigsawstack
python-dotenv